chore: make wireguard work over webrtc
