+++ browser/Makefile
@@ -1,3 +1,3 @@
 run:
-	GOOS=js GOARCH=wasm go build -o assets/tun.wasm ./wasm/
+	GOOS=js GOARCH=wasm go build -o assets/client.wasm ./client/
 	go run main.go
