+++ browser/conn/conn.go
@@ -7,6 +7,7 @@ import (
 	"errors"
 	"github.com/pion/webrtc/v3"
 	"io"
+	"log"
 	"net"
 	"time"
 )
@@ -56,6 +57,7 @@ func WrapDataChannel(rtcDataChannel *webrtc.DataChannel) (*DataChannelConn, erro
 		conn.openCond.Signal()
 	})
 	conn.dc.OnMessage(func(msg webrtc.DataChannelMessage) {
+		log.Printf("received message from data channel %d", len(msg.Data))
 		if rw != nil {
 			_, err := rw.Write(msg.Data)
 			if err != nil {
@@ -84,6 +86,7 @@ func (dc *DataChannelConn) Read(b []byte) (n int, err error) {
 
 func (dc *DataChannelConn) Write(b []byte) (n int, err error) {
 	err = dc.dc.Send(b)
+	log.Printf("writing to channel %s %v", dc.dc.Label(), dc.dc)
 	if err != nil {
 		return 0, err
 	}
@@ -165,6 +168,7 @@ func (cr ContextReadCloser) SetReadDeadline(t time.Time) error {
 }
 
 func (cr ContextReadCloser) Read(p []byte) (n int, err error) {
+	log.Printf("reading bytes ro buf of len %d", len(p))
 	done := make(chan struct{})
 	go func() {
 		n, err = cr.ReadCloser.Read(p)
